+++ packages/sdk-ts/src/client/wasm/neptune/helper.ts
@@ -1,0 +1,9 @@
+import { AssetInfo } from './types.js'
+
+export function getDenom(assetInfo: AssetInfo): string | undefined {
+  if ('native_token' in assetInfo) {
+    return assetInfo.native_token.denom
+  }
+
+  return assetInfo.token.contract_addr
+}