+++ packages/sdk-ts/src/core/modules/authz/index.ts
@@ -1,7 +1,9 @@
 import MsgGrant from './msgs/MsgGrant.js'
 import MsgRevoke from './msgs/MsgRevoke.js'
 import MsgAuthzExec from './msgs/MsgExec.js'
+import MsgGrantWithAuthorization from './msgs/MsgGrantWithAuthorization.js'
 
-export { MsgGrant, MsgRevoke, MsgAuthzExec }
+export { MsgGrant, MsgRevoke, MsgAuthzExec, MsgGrantWithAuthorization }
 export * from './utils.js'
 export * from './types.js'
+export * from './msgs/authorizations/index.js'